Move methods from session.h to session.cc
