+++ onnxruntime/hosting/http/session.cc
@@ -12,6 +12,11 @@ using tcp = boost::asio::ip::tcp;  // from <boost/asio/ip/tcp.hpp>
 
 using handler_fn = std::function<void(std::string, std::string, std::string, HttpContext&)>;
 
+HttpSession::HttpSession(std::shared_ptr<Routes> routes, tcp::socket socket)
+    : routes_(std::move(routes)), socket_(std::move(socket)), strand_(socket_.get_executor()) {
+}
+
+// Asynchronously reads the request from the socket
 void HttpSession::DoRead() {
   // Make the request empty before reading,
   // otherwise the operation behavior is undefined.
@@ -27,6 +32,7 @@ void HttpSession::DoRead() {
                            std::placeholders::_2)));
 }
 
+// Perform error checking before handing off to HandleRequest
 void HttpSession::OnRead(beast::error_code ec, std::size_t bytes_transferred) {
   boost::ignore_unused(bytes_transferred);
 
@@ -44,6 +50,7 @@ void HttpSession::OnRead(beast::error_code ec, std::size_t bytes_transferred) {
   HandleRequest(std::move(req_));
 }
 
+// After writing, make the session read another request
 void HttpSession::OnWrite(beast::error_code ec, std::size_t bytes_transferred, bool close) {
   boost::ignore_unused(bytes_transferred);
 
@@ -72,5 +79,57 @@ void HttpSession::DoClose() {
 
   // At this point the connection is closed gracefully
 }
+
+// Writes the message asynchronously back to the socket
+// Stores the pointer to the message and the class itself so that
+// They do not get destructed before the async process is finished
+// If you pass shared_from_this() are guaranteed that the life time
+// of your object will be extended to as long as the function needs it
+// Most examples in boost::asio are based on this logic
+template <class Msg>
+void HttpSession::Send(Msg&& msg) {
+  using item_type = std::remove_reference_t<decltype(msg)>;
+
+  auto ptr = std::make_shared<item_type>(std::move(msg));
+  auto self_ = shared_from_this();
+  self_->res_ = ptr;
+
+  http::async_write(self_->socket_, *ptr,
+                    net::bind_executor(strand_,
+                                       [self_, close = ptr->need_eof()](beast::error_code ec, std::size_t bytes) {
+                                         self_->OnWrite(ec, bytes, close);
+                                       }));
+}
+
+// Handle the request and hand it off to the user's function
+// Called after the session is finished reading the message
+// Should set the response before calling Send
+template <typename Body, typename Allocator>
+void HttpSession::HandleRequest(boost::beast::http::request<Body, boost::beast::http::basic_fields<Allocator> >&& req) {
+  HttpContext context{};
+  context.request = req;
+
+  std::string path = req.target().to_string();
+  std::string model_name;
+  std::string model_version;
+  std::string action;
+  handler_fn func;
+  http::status status = routes_->ParseUrl(req.method(), path, model_name, model_version, action, func);
+
+  if (http::status::ok == status) {
+    func(model_name, model_version, action, context);
+  } else {
+    http::response<http::string_body> res{status, req.version()};
+    res.set(http::field::server, BOOST_BEAST_VERSION_STRING);
+    res.set(http::field::content_type, "text/plain");
+    res.keep_alive(req.keep_alive());
+    res.body() = std::string("Something failed\n");
+    res.prepare_payload();
+    context.response = res;
+  }
+
+  return Send(std::move(context.response));
+}
+
 }  // namespace hosting
 }  // namespace onnxruntime
